% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_sample_prob.r
\name{file_sample_prop}
\alias{file_sample_prop}
\title{Proportional File Sampler}
\usage{
file_sample_prop(p, infile, outfile = tempfile(), header = TRUE,
  nskip = 0, nmax = 0, verbose = FALSE)
}
\arguments{
\item{p}{Proportion to retain; should be a numeric value between 0 and 1.}

\item{infile}{Location of the file (as a string) to be subsampled.}

\item{outfile}{Output file location (as a string).}

\item{header}{Is a header (line of column names) on the first line of the csv file?}

\item{nskip}{Number of lines to skip. If \code{header=TRUE}, then this only applies to
lines after the header.}

\item{nmax}{Max number of lines to read.  If nmax==0, then there is no read cap.}

\item{verbose}{Should linecounts of the input file and the number of lines sampled be
printed?}
}
\value{
\code{NULL}
}
\description{
Randomly sample lines from an input text file at proportion p.
}
\details{
The sampling is done in one pass of the input file, dumping lines to a
temporary file as the input is read.

If the output file (the one pointed to by the return of this function) is
"large" and to be read into memory (which isn't really appropriate for text
files in the first place!), then this strategy is probably not appropriate.
}
