% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.r
\name{file_sampler_exact}
\alias{file_sampler_exact}
\title{Exact File Sampler}
\usage{
file_sampler_exact(header, nskip, nlines, infile, outfile = tempfile())
}
\arguments{
\item{header}{Logical; indicates whether or not there is a header on the csv file.}

\item{nskip}{Number of lines to skip.  If \code{header=TRUE}, then this only
applies to lines after the header.}

\item{nlines}{The (exact) number of lines to sample from the input file.}

\item{infile}{Location of the file (as a string) to be subsampled.}

\item{outfile}{Output file.  Default is a temporary file.}
}
\description{
Randomly sample exactly \code{nlines} lines from an input text 
file.
}
\details{
The sampling is done in two passes of the input file.  First, 
the number of lines of the input file are determined by
scanning through the file as quickly as possible (i.e., it
should be completely I/O bound).  Next, an index of lines
to keep is produced by reservoir sampling.  Then finally, 
the input file is scanned again line by line with the chosen
lines dumped into a temporary file.

If the output file (the one pointed to by the return of this 
function) is "large" and to be read into memory (which isn't 
really appropriate for text files in the first place!), then 
this strategy is probably not appropriate.
}
\seealso{
\code{\link{file_sampler}}
}

