% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc.r
\name{print-wc}
\alias{print-wc}
\alias{print.wc}
\title{Print \code{wc} objects}
\usage{
\method{print}{wc}(x, ...)
}
\arguments{
\item{x}{\code{wc} object}

\item{...}{unused}
}
\description{
Printing for \code{wc()}
}
