% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_csv_sampled}
\alias{read_csv_sampled}
\title{Read Sample of CSV}
\usage{
read_csv_sampled(file, p = 0.1, header = TRUE)
}
\arguments{
\item{file}{Location of the file (as a string) to be subsampled.}

\item{p}{Proportion to retain; should be a numeric value between 0 and 1.}

\item{header}{Logical; indicates whether or not there is a header on the csv file.}
}
\value{
A dataframe, as with \code{read.csv()}.
}
\description{
The function will read (as csv) approximately p*nlines lines. So
if \code{p=.1}, then we will get roughly (probably not exactly)
10% of the data.
}
\details{
TODO
}
\examples{
\dontrun{
library(lineSampler)
file <- "/path/to/my/big.csv"
data  <- read_csv_sampled(file=file)
}
}
\seealso{
\code{\link{readLines_sampled}, \link{wc}}
}

