% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readLines_sampled}
\alias{readLines_sampled}
\title{Read Sample Lines of Text File}
\usage{
readLines_sampled(file, p = 0.1)
}
\arguments{
\item{file}{Location of the file (as a string) to be subsampled.}

\item{p}{Proportion to retain; should be a numeric value between 0 and 1.}
}
\value{
A character vector \code{readLines()}.
}
\description{
The function will read approximately p*nlines lines of a flat text
file. So if \code{p=.1}, then we will get roughly (probably not exactly)
10% of the data.
}
\details{
TODO
}
\examples{
\dontrun{
library(lineSampler)
file <- "/path/to/my/big.txt"
data  <- readLines_sampled(file=file)
}
}
\seealso{
\code{\link{read_csv_sampled}, \link{wc}}
}

