% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_csv.r
\name{sample_csv}
\alias{sample_csv}
\title{Read Sample of CSV}
\usage{
sample_csv(file, p = 0.05, header = TRUE, nskip = 0, nmax = 0,
  verbose = FALSE, ...)
}
\arguments{
\item{file}{Location of the file (as a string) to be subsampled.}

\item{p}{Proportion to retain; should be a numeric value between 0 and 1.}

\item{nskip}{Number of lines to skip.  If \code{header=TRUE}, then this only
applies to lines after the header.}

\item{nmax}{Max number of lines to read.  If nmax==0, then there is no read cap.}

\item{verbose}{Logical; indicates whether or not linecounts of the input file and the number
of lines sampled should be printed.}

\item{...}{Additional arguments passed to \code{read.csv()}.}
}
\value{
A dataframe, as with \code{read.csv()}.
}
\description{
The function will read (as csv) approximately p*nlines lines. So 
if \code{p=.1}, then we will get roughly (probably not exactly)
10% of the data.  This is the analogue of the base R function
\code{read.csv()}.
}
\details{
This function scans over the test of the input file and at each step, randomly
chooses whether or not to include the current line into a downsampled file.
Each selected line is placed in a temporary file, before being read into R
via \code{read.csv()}.  Additional arguments to this function (those other
than \code{file}, \code{p}, and \code{verbose}) are passed to \code{read.csv()},
and so if their behavior is unclear, you should examine the \code{read.csv()}
help file.

If \code{verbose=TRUE}, then something like:

\code{Read 12207 lines (0.001\%) of 12174948 line file.}

will be printed to the terminal.  This counts the header (if there is one)
as one of the lines read and as one of the lines possible.
}
\examples{
\dontrun{
library(lineSampler)
file <- "/path/to/my/big.csv"
data  <- sample_csv(file=file, sep=",")
}

}
\seealso{
\code{\link{sample_lines}, \link{wc}}
}

