% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc.r
\name{wc}
\alias{wc}
\alias{wc_w}
\alias{wc_l}
\title{Count Letters, Words, and Lines of a File}
\usage{
wc(file, chars = TRUE, words = TRUE, lines = TRUE)

wc_w(file)

wc_l(file)
}
\arguments{
\item{file}{Location of the file (as a string) from which the counts will be generated.}

\item{chars, words, lines}{Should char/word/line counts be shown? At least one of the three must be
\code{TRUE}.}
}
\value{
A list containing the requested counts.
}
\description{
See title.
}
\details{
\code{wc_l()} is a shorthand for counting only lines, similar to \code{wc -l}
in the terminal. Likewise \code{wc_w()} is analogous to \code{wc -w} for
words.
}
\examples{
library(filesampler)
file = system.file("rawdata/small.csv", package="filesampler")
data = wc(file=file)

}
