% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readers.r
\name{file_sampler}
\alias{file_sampler}
\title{File Sampler}
\usage{
file_sampler(verbose, header, nskip, nmax, p, infile, outfile = tempfile())
}
\arguments{
\item{verbose}{Logical; indicates whether or not linecounts of the input file and the number
of lines sampled should be printed.}

\item{header}{Logical; indicates whether or not there is a header on the csv file.}

\item{nskip}{Number of lines to skip.  If \code{header=TRUE}, then this only
applies to lines after the header.}

\item{p}{Proportion to retain; should be a numeric value between 0 and 1.}

\item{infile}{Location of the file (as a string) to be subsampled.}

\item{outfile}{Output file.  Default is a temporary file.}
}
\description{
Randomly sample lines from an input text file at proportion p.
}
\details{
The sampling is done in one pass of the input file, dumping
lines to a temporary file as the input is read.

If the output file (the one pointed to by the return of this
function) is "large" and to be read into memory (which isn't
really appropriate for text files in the first place!), then
this strategy is probably not appropriate.
}
\seealso{
\code{\link{file_sampler_exact}}
}

