% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_csv.r
\name{sample_csv}
\alias{sample_csv}
\title{Read Sample of CSV}
\usage{
sample_csv(
  file,
  param,
  method = "proportional",
  reader = utils::read.csv,
  header = TRUE,
  nskip = 0,
  nmax = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{file}{Location of the file (as a string) to be subsampled.}

\item{param}{The downsampling parameter. For the "proportional" method, this is the
proportion to retain and should be a numeric value between 0 and 1. For the
exact method, this is the total number of lines to read in.}

\item{method}{A string indicating the type of read method to use. Options are
"proportional" and "exact".}

\item{reader}{A function specifying the reader to use. The default is 
\code{utils::read.csv}. Other options include \code{data.table::fread()} and
\code{readr::read_csv()}.  Note the first argument of the reader should be
the file to read in and the second should be the the
\code{header}/\code{col_names} argument.  This would require writing a small
wrapper for \code{fread()}.}

\item{header}{Is a header (line of column names) on the first line of the csv file?}

\item{nskip}{Number of lines to skip. If \code{header=TRUE}, then this only applies to
lines after the header.}

\item{nmax}{Max number of lines to read. If nmax==0, then there is no read cap. Ignored
if \code{method="exact"}.}

\item{verbose}{Should linecounts of the input file and the number of lines sampled be
printed?}

\item{...}{Additional arguments passed to the csv reader.}
}
\value{
A dataframe, as with \code{read.csv()}.
}
\description{
The function will read (as csv) approximately p*nlines lines. So 
if \code{p=.1}, then we will get roughly (probably not exactly) 10% of the
data.  This is the analogue of the base R function \code{read.csv()}.
}
\details{
This function scans over the test of the input file and at each step,
randomly chooses whether or not to include the current line into a
downsampled file. Each selected line is placed in a temporary file, before
being read into R via \code{read.csv()}.  Additional arguments to this
function (those other than \code{file}, \code{p}, and \code{verbose}) are
passed to \code{read.csv()}, and so if their behavior is unclear, you should
examine the \code{read.csv()} help file.

If \code{verbose=TRUE}, then something like:

\code{Read 12207 lines (0.001\%) of 12174948 line file.}

will be printed to the terminal. This counts the header (if there is one) as
one of the lines read and as one of the lines possible.
}
\examples{
library(filesampler)
file = system.file("rawdata/small.csv", package="filesampler")

# Read in a 5\% random subsample of the rows.
data = sample_csv(file, param=.05)

# Read in 10 randomly sampled rows.
data = sample_csv(file, param=10, method="exact")

}
