% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc.r
\name{wc}
\alias{wc}
\title{Count Letters, Words, and Lines of a File}
\usage{
wc(file, chars = TRUE, words = TRUE, lines = TRUE)
}
\arguments{
\item{file}{Location of the file (as a string) from which the counts will be generated.}

\item{chars, words, lines}{Logical; show char/word/line counts be shown? At least one of
the three must be \code{TRUE}.}
}
\value{
A list containing the requested counts.
}
\description{
See title.
}
\details{
The summary statistics of the input file are determined by
scanning through the file as quickly as possible (i.e., it
should be completely I/O bound).
}
\examples{
\dontrun{
library(lineSampler)
file <- "/path/to/my/big.txt"
data  <- wc(file=file)
}

}
\seealso{
\code{\link{sample_csv}, \link{sample_lines}}
}

